+++ packages/sdk/test/ci/wasm/VerifiablePresentationRequestV1.test.ts
@@ -41,9 +41,11 @@ describe('VerifiablePresentationRequestV1', () => {
             )
             .forIdentityCredentials([0, 1, 2], (b) => b.revealAttribute(AttributeKeyString.firstName))
             .getStatements();
-        const anchor = VerifiablePresentationRequestV1.createAnchor(context, statement, 'public info');
+        const anchor = VerifiablePresentationRequestV1.createAnchor(context, statement, {
+            somePulicInfo: 'public info',
+        });
         const expectedAnchor =
-            'a4646861736858206a6f644b6d22e1647196c4fae44ffef5be554dc0edcac30518ef9624ab44de4f647479706566434344565241667075626c69636b7075626c696320696e666f6776657273696f6e01';
+            'a4646861736858206a6f644b6d22e1647196c4fae44ffef5be554dc0edcac30518ef9624ab44de4f647479706566434344565241667075626c6963a16d736f6d6550756c6963496e666f6b7075626c696320696e666f6776657273696f6e01';
         expect(Buffer.from(anchor).toString('hex')).toEqual(expectedAnchor);
 
         const roundtrip = VerifiablePresentationRequestV1.decodeAnchor(anchor);
@@ -51,7 +53,7 @@ describe('VerifiablePresentationRequestV1', () => {
             type: 'CCDVRA',
             version: 1,
             hash: VerifiablePresentationRequestV1.computeAnchorHash(context, statement),
-            public: 'public info',
+            public: { somePulicInfo: 'public info' },
         };
         expect(roundtrip).toEqual(expectedData);
     });